+++ plugin/Hydro/groundhogSurfaceFlowPathComponent.cs
@@ -28,7 +28,8 @@ namespace groundhog
         {
             pManager.AddSurfaceParameter("Surface", "S", "Base landscape form (as a surface) for the flow calculation", GH_ParamAccess.item);
             pManager.AddPointParameter("Points", "P", "Start points for the flow paths (will be projected on to the surface)", GH_ParamAccess.list);
-            pManager.AddNumberParameter("Fidelity", "F", "Amount to move for each flow iteration. Small numbers may take a long time to compute", GH_ParamAccess.item, 1000.0);
+            pManager.AddNumberParameter("Fidelity", "F", "Amount to move for each flow iteration. Small numbers may take a long time to compute. If not specified or set to 0 a (hopefully) sensible step size will be calculated.", GH_ParamAccess.item, 0);
+            pManager[2].Optional = true;
             pManager.AddIntegerParameter("Steps", "L", "A limit to the number of flow iterations. Leave unset or to 0 for an unlimited set of iterations", GH_ParamAccess.item, 0);
             pManager[3].Optional = true;
             pManager.AddBooleanParameter("Thread", "T", "Whether to multithread the solution (this can speed up long calculations)", GH_ParamAccess.item, false);
@@ -41,7 +42,7 @@ namespace groundhog
             pManager.AddCurveParameter("Flow Paths", "C", "A polyline linking each of the flow points into a path", GH_ParamAccess.list);
         }
 
-        protected override void SolveInstance(IGH_DataAccess DA)
+        protected override void GroundHogSolveInstance(IGH_DataAccess DA)
         {
             // Create holder variables for input parameters
             var FLOW_SURFACE = default(Surface);
@@ -59,8 +60,7 @@ namespace groundhog
 
             if (FLOW_FIDELITY == 0)
             {
-                AddRuntimeMessage(GH_RuntimeMessageLevel.Error, "Flow fidelity cannot be 0");
-                return;
+                FLOW_FIDELITY = FlowCalculations.getSensibleFidelity(FLOW_ORIGINS);
             }
 
             var startPoints = FLOW_ORIGINS.ToArray(); // Array for multithreading
@@ -80,28 +80,10 @@ namespace groundhog
                 for (var i = 0; i < startPoints.Length; i = i + 1)
                     allFlowPathPoints[i] = DispatchFlowPoints(FLOW_BREP, startPoints[i], FLOW_FIDELITY, FLOW_LIMIT);
 
-            var allFlowPathPointsTree = new DataTree<object>();
-            var allFlowPathCurvesList = new List<Polyline>();
-
-            for (var i = 0; i < allFlowPathPoints.Length; i++)
-            {
-                var path = new GH_Path(i);
-                // For each flow path make the polyline
-                if (allFlowPathPoints[i].Count > 1)
-                {
-                    var flowPath = new Polyline(allFlowPathPoints[i]);
-                    allFlowPathCurvesList.Add(flowPath);
-                }
-
-                // And make a branch for the list of points
-                for (var j = 0; j < allFlowPathPoints[i].Count; j++)
-                    // For each flow path point
-                    allFlowPathPointsTree.Add(allFlowPathPoints[i][j], path);
-            }
-
+            var outputs = FlowCalculations.MakeOutputs(allFlowPathPoints);
             // Assign variables to output parameters
-            DA.SetDataTree(0, allFlowPathPointsTree);
-            DA.SetDataList(1, allFlowPathCurvesList);
+            DA.SetDataTree(0, outputs.Item1);
+            DA.SetDataList(1, outputs.Item2);
         }
 
         private List<Point3d> DispatchFlowPoints(Brep FLOW_SURFACE, Point3d initialStartPoint,
@@ -140,17 +122,9 @@ namespace groundhog
 
             // Get closest point
             FLOW_SURFACE.ClosestPoint(startPoint, out closestPoint, out closestCI, out closestS, out closestT,
-                maximumDistance, out closestNormal);
-
-            // Get the vector to flow down
-            var flowVector = Vector3d.CrossProduct(Vector3d.ZAxis, closestNormal);
-            flowVector.Unitize();
-            flowVector.Reverse();
-            flowVector.Transform(Transform.Rotation(Math.PI / 2, closestNormal, closestPoint));
-
-            // Flow to the new point
-            var nextFlowPoint = Point3d.Add(closestPoint, flowVector * MOVE_DISTANCE);
-
+                                      maximumDistance, out closestNormal);
+            // Get the next point following the vector
+            var nextFlowPoint = FlowCalculations.MoveFlowPoint(closestNormal, closestPoint, MOVE_DISTANCE);
             // Need to snap back to the surface (the vector may be pointing off the edge)
             return FLOW_SURFACE.ClosestPoint(nextFlowPoint);
         }