+++ plugin/Plants/groundhogPGenericComponent.cs
@@ -10,7 +10,7 @@ namespace groundhog
     {
 
         public groundhogPGenericComponent()
-            : base("Generic Species Attributes", "Generic Species", "Output plant objects from pre-define generic types ", "Groundhog", "Flora")
+            : base("Generic Species", "PGeneric", "Output plant objects from pre-define generic types ", "Groundhog", "Flora")
         {
         }
 
@@ -31,7 +31,7 @@ namespace groundhog
             pManager.AddGenericParameter("Tree", "T", "Generic Tree (placeholder data)", GH_ParamAccess.item);
         }
 
-        protected override void SolveInstance(IGH_DataAccess DA)
+        protected override void GroundHogSolveInstance(IGH_DataAccess DA)
         {
             var csvPlantSpecies = new List<PlantSpecies>();