+++ plugin/Terrain/groundhogContourCheckPlanarityComponent.cs
@@ -11,7 +11,7 @@ namespace groundhog
     {
 
         public groundhogContourCheckPlanarityComponent()
-            : base("Contour Planarity Fix", "Contour Planarity Fix", "Checks contours are planar and corrects them if not", "Groundhog", "Terrain")
+            : base("Contour Planarity Fix", "Contour Planarity", "Checks contours are planar and corrects them if not", "Groundhog", "Terrain")
         {
         }
 
@@ -33,7 +33,7 @@ namespace groundhog
             pManager.AddCurveParameter("Fixed Contours", "FC", "Just the non-planar contours that were fixed", GH_ParamAccess.list);
         }
 
-        protected override void SolveInstance(IGH_DataAccess DA)
+        protected override void GroundHogSolveInstance(IGH_DataAccess DA)
         {
             // Create holder variables for input parameters
             var ALL_CONTOURS = new List<Curve>();